♻️ Refactor: useDetailModal 훅 분리 (#27)
